// src/game/enum/tech-type.enum.ts
export enum TechTypeEnum {
  PICKS = "PICKS",
  AXES = "AXES",
  CRUCIBLE = "CRUCIBLE",
  STORAGE = "STORAGE",
  WAGONS = "WAGONS",
  MOUNTS = "MOUNTS",
  EAGLES = "EAGLES",
  SPACEANDTIME = "SPACEANDTIME",
  WEAPONS = "WEAPONS",
  ARMORS = "ARMORS",
  ARCANOMANCY = "ARCANOMANCY",
  ESPIONAGE = "ESPIONAGE",
  LOGISTIC = "LOGISTIC",
  STRATEGY = "STRATEGY",
  EXPLORATION = "EXPLORATION",
  TARTARUS = "TARTARUS",
}